+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * (theimo1221) Minor fix to prevent uneccessary save of device settings
 * (theimo1221) Add Device Setting partial set unit test
+* (theimo1221) Add missing call for loadSettings in ioBrokerBaseDevice
 
 ## 1.3.5 (2022-11-05)